pyproject.toml
[project]
name = "thinkledger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "google-api-python-client>=2.167.0",
    "google-auth-oauthlib>=1.2.1",
    "jinja2>=3.1.6",
    "mypy>=1.15.0",
    "pip>=25.0.1",
    "plaid-python>=29.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sendgrid>=6.11.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.1",
]
